+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## DEV
+
+* **Update**: Scroll to the top on mounting
+
 ## 0.4.2
 
 * **Bugfix**: Fix a bug in the active heading selection when there are no headings